==> Complete/code/analysis/01-settings.Rmd <==
# Libraries and session info

Here we load libraries, create a folder for saving figures and define global variables. Note that this part of code is needed for the remaining code to run. 

```{r load-libraries, message=FALSE}
library(knitr)
library(bookdown)
library(VennDiagram) 
library(Rmisc)

library(tidyverse)
library(ggpubr)
library(kableExtra)
library(BayesFactor)
library(afex)

source("function/TES.R")

# automatically create a bib database for R packages
knitr::write_bib(c(
  .packages(), 'bookdown', 'knitr', 'rmarkdown'
), 'packages.bib')

# check if figure folder exists, if not, create a new one
if(!dir.exists("figure")){
  dir.create("figure")
}

# Suppress summarise info
options(dplyr.summarise.inform = FALSE)

# set the theme to theme_bw for all ggplot2 figures
theme_set(theme_bw())

# use colorblind friendly colors
color_values <- c("#56B4E9", "#E69F00")
```


```{r session-info}
sessionInfo()
```
