==> Complete/code/analysis/02-process_data.Rmd <==
# Process data

The raw data on the round level cannot be shared as they are proprietary. We nevertheless provide the code that we used to process the raw data to get the player-level summary data. The summary data are in the `data/processed` folder and can be used to reproduce the results in the paper.

```{r load-raw-data}
#### load data from high-risk players ####

# get all folders where data from high-risk players are saved
folders <- list.dirs("../../data/raw/data_addict_players", full.names = TRUE, recursive = FALSE)

# get all file names
files <- paste0(folders, "/rounds_data.csv")

# read csv files
d_high_risk <- files %>%
  map_dfr(~read_csv(.x, skip = 1,
                    col_names = c("id_player", "level","id_session", "id_round", "stake", "win", 
                                  "rt_start", "rt1", "rt2", "rt3", "rt4", "rt5", "rt6", "rt7", "rt8", 
                                  "rt9", "rt10", "rt11", "rt12", "bonus"),
                    col_types =  cols(.default = col_character())))

# load demographic data (age and gender)
d_high_risk_demo <- read_delim("../../data/raw/meta_data_addict_players.csv", delim = ";",
                               col_types = cols(.default = col_character()))

# add demographic information to the main data frame
d_high_risk <- d_high_risk %>%
  left_join(d_high_risk_demo, by = c("id_player" = "id"))

#### load data from low-risk players ####

# get all folders where data from low-risk players are saved
folders <- list.dirs("../../data/raw/data_non_addict_players", full.names = TRUE, recursive = FALSE)

# get all file names
files <- paste0(folders, "/rounds_data.csv")

# read csv files
d_low_risk <- files %>%
  map_dfr(~read_csv(.x, skip = 1,
                    col_names = c("id_player", "id_session", "id_round", "stake", "win", 
                                  "rt_start", "rt1", "rt2", "rt3", "rt4", "rt5", "rt6", "rt7", "rt8", 
                                  "rt9", "rt10", "rt11", "rt12", "bonus"),
                    col_types =  cols(.default = col_character())))


# load demographic data (age and gender)
d_low_risk_demo <- read_delim("../../data/raw/meta_data_non_addict_players.csv", delim = ";",
                              col_types = cols(.default = col_character()))

# add demographic information to the main data frame
d_low_risk <- d_low_risk %>%
  left_join(d_low_risk_demo, by = c("id_player" = "id"))

#### quality checks ####

# check if the two groups consist of different players
high_risk_players <- unique(d_high_risk$id_player)
low_risk_players <- unique(d_low_risk$id_player)

# all(high_risk_players %in% low_risk_players == FALSE)
# all(low_risk_players %in% high_risk_players == FALSE)

# both commands above return TRUE, showing that 
# the two groups have different anonymous player IDs.
# Thus, the two groups indeed consist of different players.

#### combine data ####

d_low_risk <- d_low_risk %>%
  mutate(
    # add group name,
    group = "Low-Risk",
    # low-risk players all have a risk level of 0
    level = 0
  ) %>%
  select(group, level, everything())

d_high_risk <- d_high_risk %>%
  mutate(
    # add group name,
    group = "High-Risk"
  ) %>%
  select(group, level, everything())

# combine data from both groups
d <- rbind(d_high_risk, d_low_risk)

# remove the original two data frames
remove(d_high_risk)
remove(d_low_risk)

#### clean data ####

# change player id, session id and round id into integers
d <- d %>%
  # each player gets a unique number, starting from 1
  mutate(id_player = match(id_player, unique(id_player))) %>%
  # within each player, each session gets a unique number, starting from 1 
  group_by(id_player) %>%
  mutate(id_session = match(id_session, unique(id_session))) %>%
  # each round within each session gets a unique number, starting from 1
  group_by(id_player, id_session) %>%
  mutate(id_round = row_number()) %>%
  ungroup()

# change some variables (stake, win, RTs etc.) into numeric
d <- d %>%
  mutate(
    across(stake:rt12, as.numeric),
    across(c(level, age), as.numeric)
  )

# create new variables for the stake,  win amount on the previous round, 
# and whether there is a bonus game or not on the previous round.
d <- d %>%
  group_by(id_player, id_session) %>%
  mutate(
    prev_stake = lag(stake),
    prev_win = lag(win),
    prev_bonus = lag(bonus)
  ) %>%
  ungroup()

```

## Player and play behavior

### descriptive_data.csv

First, we process the data to get descriptive data on the players and their general play behaviors.

The `descriptive_data.csv` file contains the following variables.

* group: the group to which a player belongs, High-Risk or Low-Risk.
* id_player: the anonymous id of a player.
* age: age of a player.
* gender: gender of a player, Female or Male.
* risk_0: whether the player receives a risk level of 0 in the data, 1 = yes, 0 = no.
* risk_3: whether the player receives a risk level of 3 in the data, 1 = yes, 0 = no.
* risk_4: whether the player receives a risk level of 4 in the data, 1 = yes, 0 = no.
* risk_5: whether the player receives a risk level of 5 in the data, 1 = yes, 0 = no.
* n_session: the total number of sessions played by each player.
* n_round_total: the total number of rounds played by each player.
* n_round_mean: the mean number of rounds played per session.
* n_round_median: the median number of rounds played per session.
* bet_max: the maximum bet size across all rounds, in euro.
* bet_min: the minimum bet size across all rounds, in euro.
* bet_mean: the mean bet size across all rounds, in euro.
* bet_median: the median bet size across all rounds, in euro.
* win_prob: the probability of winning (%) across all rounds.
* win_max: the maximum amount of win on rounds when players win, in euro. Win defined as the win amount minus the stake amount.
* win_min: the minimum amount of win on rounds when players win, in euro.
* win_mean: the mean amount of win on rounds when players win, in euro.
* win_median:  the median amount of win on rounds when players win, in euro.
* loss_max: the maximum amount of loss on rounds when players lose, in euro. Loss amount equals the stake amount.
* loss_min: the minimum amount of loss on rounds when players lose, in euro.
* loss_mean: the mean amount of loss on rounds when players lose, in euro.
* loss_median: the median amount of loss on rounds when players lose, in euro.
* total_spent: the total amount of money spent (i.e., lost), in euro. A negative value indicates that players overall won money.

```{r descriptive_data}
# get the age and gender of each player
age_gender <- d %>%
  group_by(id_player) %>%
  sample_n(1) %>%
  mutate(gender = recode(gender, F = "Female", M = "Male")) %>%
  select(group, id_player, age, gender)

# get the risk level(s) of each player
risk_level_0 <- d %>%
  filter(level == 0) %>%
  .$id_player %>% unique

risk_level_3 <- d %>%
  filter(level == 3) %>%
  .$id_player %>% unique

risk_level_4 <- d %>%
  filter(level == 4) %>%
  .$id_player %>% unique

risk_level_5 <- d %>%
  filter(level == 5) %>%
  .$id_player %>% unique

risk_levels <- age_gender %>%
  mutate(
    risk_0 = ifelse(id_player %in% risk_level_0, 1, 0),
    risk_3 = ifelse(id_player %in% risk_level_3, 1, 0),
    risk_4 = ifelse(id_player %in% risk_level_4, 1, 0),
    risk_5 = ifelse(id_player %in% risk_level_5, 1, 0)
  ) %>%
  select(id_player, risk_0, risk_3, risk_4, risk_5)

# count the number of sessions and rounds for each player
rounds <- d %>%
  # count the number of rounds in each session
  count(id_player, id_session) %>%
  # count for each player:
  group_by(id_player) %>%
  summarize(
    n_session = n(), # the total number of sessions
    n_round_total = sum(n), # the total number of rounds (all sessions combined)
    n_round_mean =  mean(n), # the mean number of rounds per session
    n_round_median = median(n) # the median number of rounds per session
  ) %>%
  select(id_player, n_session, n_round_total, n_round_mean, n_round_median)

# for each player, calculate the max, min, mean and median of stakes across all rounds
bets <- d %>%
  group_by(id_player) %>%
  summarize(
    bet_max = max(stake),
    bet_min = min(stake),
    bet_mean = mean(stake),
    bet_median = median(stake)
  ) %>%
  select(id_player, bet_max, bet_min, bet_mean, bet_median)

# for each player, calculate their winning probability, the min, max, mean and median win amount
win_prob <- d %>%
  group_by(id_player) %>%
  summarize(win_prob = mean(win > 0) * 100) %>%
  select(id_player, win_prob)

wins <- d %>%
  # select winning rounds
  filter(win > 0) %>%
  # for each player, calculate the max, min, mean and median win amount
  # the real win amount is the win amount displayed by the game, minus the stake
  mutate(win = win - stake) %>%
  group_by(id_player) %>%
  summarize(
    win_max = max(win),
    win_min = min(win),
    win_mean = mean(win),
    win_median = median(win)
  ) %>%
  select(id_player, win_max, win_min, win_mean, win_median)

# on the rounds where players lost
# calculate their max, min, mean, median and total loss amount
losses <- d %>%
  # select losing rounds
  filter(win == 0) %>%
  # loss the same as the stake
  mutate(loss = stake) %>%
  group_by(id_player) %>%
  summarize(
    loss_max = max(loss),
    loss_min = min(loss),
    loss_mean = mean(loss),
    loss_median = median(loss),
  ) %>%
  select(id_player, loss_max, loss_min, loss_mean, loss_median)

# for each player, calculate the total spent
total_spent <- d %>%
  mutate(spent = stake - win) %>%
  group_by(id_player) %>%
  summarize(total_spent = sum(spent))

# put all data frames together
d_descriptive <- age_gender %>%
  full_join(risk_levels, by = "id_player") %>%
  full_join(rounds, by = "id_player") %>%
  full_join(bets, by = "id_player") %>%
  full_join(win_prob, by = "id_player") %>%
  full_join(wins, by = "id_player") %>%
  full_join(losses, by = "id_player") %>%
  full_join(total_spent, by = "id_player")

# save the data frames as a csv file
write_csv(d_descriptive, "../../data/processed/descriptive_data.csv")

# remove data frames
remove(age_gender, risk_levels, rounds, bets, win_prob, wins, losses, total_spent, d_descriptive)
```

### bet_counts.csv

On each round, players can choose one of ten different bet sizes (from 0.25 euro to 20 euro). To characterize the betting patterns, for each player we count the number of rounds where they chose a certain bet size. The data are saved in the `bet_counts.csv` file, which contains the following variables.

* group: the group to which a player belongs, High-Risk or Low-Risk.
* id_player: the anonymous id of a player.
* stake: the amount of money a player bets on a round, from 0.25 till 20, in euro.
* n_round: the number of rounds where a player chose a certain bet size.
* n_round_total: the total number of rounds played by each player.
* prop: the relative proportion of choosing a certain bet size for each player, from 0 (%) till 100 (%).

```{r bet-patterns}
# get all the unique levels of bet sizes
bet_sizes <- unique(d$stake)
bet_sizes <- sort(bet_sizes, decreasing = F)

# for each player, count how often each bet size is selected
bets_count <- d %>%
  # turn the variable stake into a factor with all ten 10 possible levels
  mutate(stake = factor(stake, levels = bet_sizes)) %>%
  # for each player, count the number of rounds with a certain stake size
  count(group, id_player, stake, .drop = F, name = "n_round") %>%
  # calculate the total number of rounds for each player
  # and based on that, the relative proportion of choosing a certain stake size
  group_by(group, id_player) %>%
  mutate(
    n_round_total = sum(n_round), # the total number of rounds
    prop = n_round/n_round_total * 100  # the relative proportion of choosing a certain bet
  ) %>%
  ungroup()

# save the data frame as a csv file
write_csv(bets_count, "../../data/processed/bets_count.csv")

# remove data frame(s)
remove(bets_count)
```


## When to stop

### stop.csv

After each round, players can choose either to continue playing, or stop. Here we prepare data on the probability of ending a session after a win or a loss. The data are saved in the `stop.csv` file, which contains the following variables.

* group: the group to which a player belongs, High-Risk or Low-Risk.
* id_player: the anonymous id of a player.
* outcome: the outcome of a round, win or loss.
* total_count: the total number of rounds with a particular outcome (win or loss).
* stop_count: the total number of rounds after which a player decided to end a session.
* stop_prob: the probability (%) of ending a session after a particular outcome. This variable is computed by dividing stop_count by total_count and then multiplied by 100.
* stop_overall: the overall probability (%) of ending a session regardless of the outcome of a round. This variable is computed by dividing the number of rounds at the end of a session by the total number of rounds (all sessions combined), and then multiplied by 100.


```{r when to stop}
# for each player, compute their overall probability of stopping
stop_prob <- d %>%
  group_by(group, id_player, id_session) %>%
  mutate(
    # if a certain round has the largest round id in a session
    # it means that round is the last one, and player ended the session afterwards
    round_pos = ifelse(id_round == max(id_round), "end", "in")
  ) %>%
  group_by(group, id_player) %>%
  summarise(
    total_count = n(),
    stop_count = sum(round_pos == "end"),
    stop_overall = stop_count/total_count * 100
  )

# for each player, compute the probability of stopping after a win and a loss
stop_prob_outcome <- d %>%
  group_by(group, id_player, id_session) %>%
  mutate(
    # if a certain round has the largest id in a session
    # it means that round is the last one, and player ended the session afterwards
    round_pos = ifelse(id_round == max(id_round), "end", "in"),
    # determine whether each round is a win or a loss
    outcome = ifelse(win > 0, "win", "loss")
  ) %>%
  group_by(group, id_player, outcome) %>%
  summarise(
    total_count = n(), # the total number of winning or losing rounds
    stop_count = sum(round_pos == "end"), # the number of winning or losing rounds at session end
    stop_prob = stop_count/total_count * 100 # the probability of stopping after winning or losing
  )

# combine the overall probability of stopping and 
# the probability of stopping after winning or losing
stop <- stop_prob %>%
  select(group, id_player, stop_overall) %>%
  full_join(stop_prob_outcome, by = c("group", "id_player")) %>%
  select(group, id_player, outcome, total_count, stop_count,
         stop_prob, stop_overall)

# save the data frame as a csv file
write_csv(stop, "../../data/processed/stop.csv")

# remove data frame(s)
remove(stop, stop_prob, stop_prob_outcome)
```


### stop_breaks.csv

Players might take a break without logging off. To take such breaks in play into account, we assume that whenever players took more than 10 minutes to make a move, it means that they had taken a break. A new session starts after such a break. The data are saved in the `stop_breaks.csv` file, which contains the following variables.

* group: the group to which a player belongs, High-Risk or Low-Risk.
* id_player: the anonymous id of a player.
* outcome: the outcome of a round, win or loss.
* total_count: the total number of rounds with a particular outcome (win or loss).
* stop_count: the total number of rounds after which a player decided to end a session.
* stop_prob: the probability (%) of ending a session after a particular outcome. This value is computed by dividing stop_count by total_count and then multiplied by 100.
* stop_overall: the overall probability (%) of ending a session regardless of the outcome of a round. This variable is computed by dividing the number of rounds at the end of a session by the total number of rounds (all sessions combined), and then multiplied by 100.


```{r when to stop with breaks in play}
# when any one of the 13 responses took longer than 10 minutes 
# (i.e., 10 * 60 * 1000 milliseconds), a break occurs
break_duration <- 10*60*1000

d <- d %>%
  rowwise() %>%
  mutate(
    breaks = ifelse(
      any(rt_start > break_duration,
          rt1 > break_duration,
          rt2 > break_duration,
          rt3 > break_duration,
          rt4 > break_duration,
          rt5 > break_duration,
          rt6 > break_duration,
          rt7 > break_duration,
          rt8 > break_duration,
          rt9 > break_duration,
          rt10 > break_duration,
          rt11 > break_duration,
          rt12 > break_duration), 
      1,
      0
    )
  )

# define new sessions by taking breaks into account
d <- d %>%
  ungroup() %>%
  mutate(
    break_count = cumsum(breaks),
    new_id_session = paste(id_session, break_count, sep = "_")
  ) %>%
  # within each player, each session gets a unique number, starting from 1 
  group_by(id_player) %>%
  mutate(new_id_session = match(new_id_session, unique(new_id_session)))

# check the number of sessions when breaks are accounted for
sessions_count <- d %>%
  group_by(id_player) %>%
  summarize(
    n_old_session = n_distinct(id_session),
    n_new_session = n_distinct(new_id_session)
  ) %>%
  mutate(diff = n_new_session - n_old_session)

# now compute the probabilities of stopping 

# for each player, compute their overall probability of stopping
stop_prob <- d %>%
  group_by(group, id_player, new_id_session) %>%
  mutate(
    # if a certain round has the largest id in a session
    # it means that round is the last one, and player ended the session afterwards
    round_pos = ifelse(id_round == max(id_round), "end", "in")
  ) %>%
  group_by(group, id_player) %>%
  summarise(
    total_count = n(),
    stop_count = sum(round_pos == "end"),
    stop_overall = stop_count/total_count * 100
  )

# for each player, compute the probability of stopping after a win or a loss
stop_prob_outcome <- d %>%
  group_by(group, id_player, new_id_session) %>%
  mutate(
    # if a certain round has the largest id in a session
    # it means that round is the last one, and player ended the session afterwards
    round_pos = ifelse(id_round == max(id_round), "end", "in"),
    # determine if each round is a win or a loss
    outcome = ifelse(win > 0, "win", "loss")
  ) %>%
  group_by(group, id_player, outcome) %>%
  summarise(
    total_count = n(),
    stop_count = sum(round_pos == "end"),
    stop_prob = stop_count/total_count * 100
  )

# add the overall probability of stopping as a column
stop_breaks <- stop_prob %>%
  select(group, id_player, stop_overall) %>%
  full_join(stop_prob_outcome, by = c("group", "id_player")) %>%
  select(group, id_player, outcome, total_count, stop_count,
         stop_prob, stop_overall)

# save the data frame as a csv file
write_csv(stop_breaks, "../../data/processed/stop_breaks.csv")

# remove data frame(s)
remove(stop_breaks, stop_prob, stop_prob_outcome)

```


## Stake change

Players can change their stakes between rounds. We explore whether the probability of changing stakes is influenced by prior outcome, and the change in stake amount as a function of prior outcome.

### stake_change.csv

Players can choose among 10 different levels of stake sizes (from 0.25 euro to 20 euro). Since we are interested in whether wins and losses may change the betting behavior, we exclude the first round of each session (i.e., these rounds have no prior outcomes). For the remaining data, we calculate the number of rounds that follow a win or a loss, and the probability of changing stakes.

We also calculate the average change in stakes (in euro; a positive change score indicates increasing stake, while a negative change score indicates decreasing stake). Since players can only select one from ten levels of stakes, as a second measure of stake change, we transform the stake amounts into stake levels, with 1 being the lowest level of stake (0.25 euro) and 10 being the highest level of stake (20 euro). A change score is then calculated based on the levels (rather than the stake amounts) of stakes. We calculate the average change in stake sizes and stake levels both for all rounds, and only on rounds where players did change stakes.

The following variables are calculated and saved in the `stake_change.csv` file.

* group: the group to which a player belongs, High-Risk or Low-Risk.
* id_player: the anonymous id of a player.
* prev_outcome: the outcome on the previous round, win (win amount > 0) or loss (win amount = 0).
* n_round: the total number of rounds following a win or a loss.
* n_round_change: the number of rounds where a player changed the stake.
* stake_change_prop: the proportion of rounds where players changed the stake (%).
* stake_change_size_overall: the average change in stake sizes in all rounds. A positive value indicates an increase in stake (euro).
* stake_change_level_overall: the average change in stake levels in all rounds. A positive value indicates an increase in stake level.
* stake_change_size: the average change in stake sizes when players did change the stake. In other words, rounds where players did not change the stake are excluded.
* stake_change_level: the average change in stake levels when players did change the stake. In other words, rounds where players did not change the stake are excluded.

```{r stake_change}
# get all the unique levels of bet sizes
bet_sizes <- unique(d$stake)
bet_sizes <- sort(bet_sizes, decreasing = F)

stake_change <- d %>%
  mutate(
    # determine if the previous round is a win or a loss
    prev_outcome = ifelse(prev_win > 0, "win", "loss"),
    # create a new variable for stake level
    stake_level = factor(stake, levels = bet_sizes),
    stake_level = as.numeric(stake_level),
    # create a new variable for the previous stake
    prev_stake_level = lag(stake_level),
    # calculate the change in stakes between two rounds
    # a positive value indicates an increase in stake
    stake_change_size = stake - prev_stake,
    # do the same for the change in stake level
    stake_change_level = stake_level - prev_stake_level
  ) %>%
  filter(!is.na(prev_outcome)) %>%
  ungroup()

# calculate the probability of changing stakes and the average change
# in stake sizes and stake levels with all rounds included
stake_change_prop <- stake_change %>%
  group_by(group, id_player, prev_outcome) %>%
  summarize(
    # the total number of rounds across all sessions
    n_round = n(),
    # the total number of rounds when stakes did change
    n_round_change = sum(stake_change_size != 0),
    # probability of changing stakes
    stake_change_prop = mean(stake_change_size != 0) * 100,
    # the overall change in stake sizes and levels
    stake_change_size_overall = mean(stake_change_size),
    stake_change_level_overall = mean(stake_change_level)
  ) %>%
  ungroup()

# calculate the changes in stakes when players did change stakes
stake_change_sizes <- stake_change %>%
  # only include rounds where players did change stakes
  filter(stake_change_size != 0) %>%
  group_by(group, id_player, prev_outcome) %>%
  summarize(
    stake_change_size = mean(stake_change_size),
    stake_change_level = mean(stake_change_level),
  ) %>%
  ungroup()

# combine both data frames
stake_change_combined <- stake_change_prop %>%
  full_join(stake_change_sizes, by = c("group", "id_player", "prev_outcome"))

# save the data frame as a csv file
write_csv(stake_change_combined, "../../data/processed/stake_change.csv")

remove(stake_change, stake_change_prop, stake_change_sizes, stake_change_combined)
```

## Speed of play

Next we examine the speed of play, once players have decided to continue playing. For this analysis, we focus on the RT of putting in the first column as an indicator of the speed of play. For this analysis, we further exclude rounds where the RT of putting in the first column was at least 5000 milliseconds. RTs below 0 (potentially a recording error) are also excluded. 

### rt.csv

```{r rt-data-exclusion}
# number of rounds before exclusion
before <- nrow(d)

d_after <- d %>%
  mutate(rt = rt1) %>%
  filter(
    # exclude the first round of each session
    id_round != 1, 
    # only keep trials where the RT of putting in the first column
    # is between 0 and 5000 milliseconds
    rt < 5000, rt >= 0 
  ) 

# calculate number of rounds after the exclusion
after <- nrow(d_after)

# calculate the proportion of data exclusion
removed <- (before - after)/before * 100
```

Players in total played `r before` rounds. `r before - after` rounds were excluded after applying the exclusion criteria as listed above (`r round(removed, 1)`\% of all rounds). For the remaining rounds, we calculate the number of rounds, and the mean RTs following wins and losses respectively. Since players differed in how quickly they generally responded, to control for the potential effects of general responding speeds, we further calculate the z score of each RT within each player. The mean RT z scores are further calculated for each player. The processed data are saved in the `rt.csv` file. The following variables are included.

* group: the group to which a player belongs, High-Risk or Low-Risk.
* id_player: the anonymous id of a player.
* prev_outcome: the outcome on the previous trial, win (win amount > 0) or loss (win amount = 0).
* n_round: the number of rounds following a win or a loss.
* rt_mean: the mean response time (of putting in the first column) after a win or a loss, in milliseconds.
* rt_z_mean: the mean of response time z score (of putting in the first column) after a win or a loss.

```{r rt-data}
rt_data <- d_after %>%
  # calculate the following variables for each player
  group_by(group, id_player) %>%
  mutate(
    # standardize the RT data within each player
    rt_z = scale(rt),
    # determine whether a previous round is a win or a loss
    prev_outcome = ifelse(prev_win > 0, "win", "loss")
  ) %>%
  # for rounds following wins and losses, calculate the total round number
  # and the mean RTs (raw scores as well as z scores)
  group_by(group, id_player, prev_outcome) %>%
  summarize(
    n_round = n(), # number of rounds
    rt_mean = mean(rt), # the mean rt 
    rt_z_mean = mean(rt_z) # the mean rt z score
  )

# save the data frame as a csv file
write_csv(rt_data, "../../data/processed/rt.csv")

remove(d_after, rt_data, d)
```
