==> Complete/code/analysis/03-descriptives.Rmd <==
# Descriptives

In this part, we provide some descriptive information of both the players and their general playing behavior in the game.

```{r descriptives-load-data}
demo <- read_csv("../../data/processed/descriptive_data.csv")
sessions_count <- read_csv("../../data/processed/sessions_count.csv")
bets_count <- read_csv("../../data/processed/bets_count.csv")
```

## Player characteristics

### Age and gender.

First, we calculated the number of players (males and females) and the descriptive information of age in each group (see Table \@ref(tab:demographics)). We further made an age by gender pyramid for each group (see Figure \@ref(fig:age-pyramids)).

```{r demographics, results='asis'}
# the total number of players; both groups combined
n_total <- nrow(demo)

# the number of players in the high-risk and the low-risk group
n_HR    <- sum(demo$group == "High-Risk")
n_LR    <- sum(demo$group == "Low-Risk")

# count the number of males and females in each group
gender_count <- demo %>%
  count(group, gender) %>%
  pivot_wider(id_cols = group, names_from = gender, values_from = n) %>%
  mutate(Total = Female + Male)

# calculate the descriptive information of age in each group
age_des <- demo %>%
  group_by(group) %>%
  summarize(across(age, list(mean = mean, median = median, sd = sd, min = min, max = max))) %>%
  mutate(age_mean = round(age_mean, 1), age_sd = round(age_sd, 1))

# combine gender count and age descriptive
demo_summary <- gender_count %>%
  inner_join(age_des, by = "group")

# show table
kable(demo_summary, caption = "Demographic information",
      booktabs = T, align = "c")
```

(ref:age-pyramids-cap) Age by gender pyramids for the high-risk and the low-risk group.

```{r age-pyramids, fig.cap='(ref:age-pyramids-cap)', fig.height=6, fig.width=8}
# age pyramids for both groups
age_count <- demo %>%
  mutate(
    # in the sample age ranges from 22 to 89. 
    # divide players into age groups from 20-90, with bin width of 5 years.
    age_range = cut(age, breaks = seq(20, 90, 5))
  ) %>%
  # count the number of males and females in each age group
  count(group, age_range, gender)

age_pyramids <- ggplot(age_count, aes(age_range, n, fill=gender)) +
  # plot bars for males
  geom_col(data=filter(age_count, gender =="Male")) +
  # add the counts next to the bars
  geom_text(data=filter(age_count, gender =="Male"), aes(x=age_range, y = n, label = n), nudge_y = 20) +
  # plot bars for females, but flip it to the other side
  geom_col(data=filter(age_count, gender=="Female"), aes(y = n*(-1))) + 
  # add the counts next to the bars
  geom_text(data=filter(age_count, gender =="Female"), aes(x=age_range, y = n*(-1), label = n), nudge_y = -20) +
  # for high-risk and low-risk group separately
  facet_wrap(~group) +
  # use colorblind-friendly color
  scale_fill_manual(values = color_values) + 
  scale_color_manual(values = color_values) + 
  labs(x = "Age Range (years old)", y = "Number of Players", fill = "Gender") +
  theme(legend.position = "top", legend.direction = "horizontal") + 
  # adjust the tick labels on the y axis
  scale_y_continuous(limits = c(-240, 240), breaks=seq(-240,240,40), labels=abs(seq(-240,240,40))) +
  # flip the x and y axis
  coord_flip()

age_pyramids

# save plots
ggsave("figure/age_pyramid.pdf", age_pyramids, width = 8, height = 6)

```

### Risk levels

All players in the low-risk group have a risk level of 0. Players in the high-risk group, however, may have more than one risk level (3, 4, or 5). This is because the risk level a certain player received is determined by the amount of risk indicators they exhibited. The more risk indicators a player exhibited, the higher the risk level was. Since players' behavior can vary over time, the risk level a particular player receives can also vary. We therefore made a Venn diagram to show the number of players with each risk level(s) in the high-risk group (Figure \@ref(fig:risk-level-venn-diagram)).

(ref:venn-diagram-cap) Venn diagram for the risk levels of players in the high-risk group. The overlapping regions of the circles indicate the numbers of players with corresponding two or three risk levels. The non-overlapping regions of the circles indicate the numbers of players with one risk level.

```{r risk-level-venn-diagram, fig.cap = '(ref:venn-diagram-cap)', fig.align="center"}
# get the player ids for each risk level
risk_level_3 <- demo %>%
  filter(risk_3 == 1) %>%
  .$id_player %>% unique

risk_level_4 <- demo %>%
  filter(risk_4 == 1) %>%
  .$id_player %>% unique

risk_level_5 <- demo %>%
  filter(risk_5 == 1) %>%
  .$id_player %>% unique

# draw a Venn diagram, code from https://www.r-graph-gallery.com/14-venn-diagramm.html

# Chart
VennDiagram::venn.diagram(
  x = list(risk_level_3, risk_level_4, risk_level_5),
  category.names = c("Level 3" , "Level 4" , "Level 5"),
  filename = 'figure/risk_levels.png',
  output=TRUE,
  
  # Output features
  imagetype="png" ,
  height = 480 , 
  width = 480 , 
  resolution = 300,
  compression = "lzw",
  
  # Circles
  lwd = 1,
  col=c("#440154ff", '#21908dff', '#fde725ff'),
  fill = c(alpha("#440154ff", 0.3), alpha('#21908dff', 0.3), alpha('#fde725ff', 0.3)),
  
  # Numbers
  cex = .6,
  fontface = "bold",
  fontfamily = "sans",
  
  # Set names
  cat.cex = 0.6,
  cat.fontface = "plain",
  cat.default.pos = "outer",
  cat.pos = c(-27, 27, 135),
  cat.dist = c(0.055, 0.055, 0.085),
  cat.fontfamily = "sans",
  rotation = 1
)

# include image in the graphics
include_graphics('figure/risk_levels.png')
```

## Play behavior characteristics

### Number of sessions and rounds

To provide some descriptive information on the total numbers of sessions players played, as well as the numbers of rounds played in one session, we count the number of players with certain numbers of sessions in the data set. We also count the numbers of sessions with certain numbers of rounds (combining players from the same group). Figure \@ref(fig:counts-plot) shows the results. The majority of the players, especially those in the low-risk group, tended to play relatively few sessions (Figure \@ref(fig:counts-plot), top panel). Players in the high-risk group though did tend to play multiple sessions. Furthermore, most of the sessions tended to be relatively short (Figure \@ref(fig:counts-plot), bottom panel), but longer playing sessions also occurred, especially for the high-risk group.

(ref:counts-plot-cap) Histograms of the numbers of players with certain numbers of sessions (top) and the numbers of sessions with certain numbers of rounds (bottom, combined across all players from the same group).

```{r counts-plot, fig.cap='(ref:counts-plot-cap)', fig.height=7, fig.width=7}
# count the number of players with certain number of sessions
players_with_certain_sessions <- demo %>%
  # divide the total number of sessions into bins
  # note that the scale is transformed so that larger values are compressed
  mutate(n_session_bin = cut(n_session, 
                             breaks = c(0:10, (2:10)*10, (2:10)*100, (2:5)*1000)),
         n_session_label = cut(n_session, 
                             breaks = c(0:10, (2:10)*10, (2:10)*100, (2:5)*1000),
                             labels = 1:32)
  ) %>%
  # count the number of players with certain numbers of sessions
  count(group, n_session_bin, n_session_label) %>%
  mutate(n_session_label = as.numeric(n_session_label)) 
  
# make a histogram for the total number of sessions
session_hist <- ggplot(players_with_certain_sessions, aes(n_session_label, n, fill = group, color = group)) +
  geom_bar(stat = "identity", alpha = 0.4, position = "identity") +
  scale_x_continuous(breaks = 0:32 + 0.5, # note that +0.5 to move the bars horizontally
                     labels = c(0:10, (2:10)*10, (2:10)*100, (2:5)*1000),
                     limits = c(0, 34)) +
  scale_y_continuous(breaks = (0:9) * 50,
                     labels = (0:9) * 50,
                     limits = c(0, 450)) +
  labs(x = "Number of Sessions", y = "Number of Players", fill = "Group", color = "Group") +
  # use colorblind-friendly color
  scale_fill_manual(values = color_values) + 
  scale_color_manual(values = color_values) + 
  theme(legend.position = "top", legend.direction = "horizontal", 
        panel.grid.minor = element_blank(),
        axis.text.x=element_text(angle = -90, vjust = 0.5))

# get the number of sessions with certain number of rounds
sessions_with_certain_rounds <- sessions_count %>%
  # divide the number of sessions in each bin by 100, for ease of plotting
  mutate(n_session = n_session / 100)
  
# make a histogram for the number of rounds
round_hist <- ggplot(sessions_with_certain_rounds, aes(n_round_label, n_session, fill = group, color = group)) +
  geom_bar(stat = "identity", alpha = 0.4, position = "identity") +
  scale_x_continuous(breaks = c(0:34) + 0.5,
                     labels = c((0:20)*10, (3:10)*100, (2:7)*1000),
                     limits = c(0, 35)) +
  scale_y_continuous(breaks = (0:16) * 10,
                     labels = (0:16) * 10,
                     limits = c(0, 160)) +
  labs(x = "Number of Rounds", y = "Number of Sessions / 100", fill = "Group", color = "Group") +
  # use colorblind-friendly color
  scale_fill_manual(values= color_values) + 
  scale_color_manual(values= color_values) + 
  theme(legend.position = "top", legend.direction = "horizontal",
        panel.grid.minor = element_blank(),
        axis.text.x=element_text(angle = -90, vjust = 0.5))

# save two plots in one figure
p <- ggarrange(session_hist, round_hist, ncol = 1,
               common.legend = TRUE, legend="top")

p

ggsave("figure/session_round_count.pdf", p, width = 7, height = 7)
```

For each player, we calculate the total number of sessions played, the total number of rounds played, the mean number of rounds played per session, and the median number of rounds played per session. We then compare the numbers of sessions and rounds between the two groups. As can be seen from Table \@ref(tab:rounds-compare), players in the high-risk group overall played more sessions and rounds than those in the low-risk group. They also tended to have longer sessions (i.e., more rounds per session) than the low-risk group.

```{r rounds-compare, results="asis"}
# calculate the total numbers of sessions and rounds across all players
n_session_total <- sum(demo$n_session)
n_round_total   <- sum(demo$n_round_total)

# compare the two groups
rounds_HR <- demo %>% filter(group == "High-Risk")
rounds_LR <- demo %>% filter(group == "Low-Risk")

session_comp      <- TES(rounds_HR$n_session, rounds_LR$n_session, paired = F, conf = 0.95)
round_total_comp  <- TES(rounds_HR$n_round_total, rounds_LR$n_round_total, paired = F, conf = 0.95)
round_mean_comp   <- TES(rounds_HR$n_round_mean, rounds_LR$n_round_mean, paired = F, conf = 0.95)
round_median_comp <- TES(rounds_HR$n_round_median, rounds_LR$n_round_median, paired = F, conf = 0.95)

session_comp <- session_comp %>% 
  mutate(comp = "Session number") %>%
  select(comp, everything())

round_total_comp <- round_total_comp %>% 
  mutate(comp = "Round number") %>%
  select(comp, everything())

round_mean_comp <- round_mean_comp %>%
  mutate(comp = "Mean round number") %>%
  select(comp, everything())

round_median_comp <- round_median_comp %>%
  mutate(comp = "Median round number") %>%
  select(comp, everything())

# combine and format the results
count_comp <- rbind(session_comp, round_total_comp, round_mean_comp, round_median_comp)

comp_formatted <- count_comp %>%
  mutate(
    across(c(mean_x, sd_x, mean_y, sd_y, diff, lowerCI, upperCI), ~round(.x)),
    across(c(df, t, logBF), ~round(.x, digits = 1)),
    across(c(d, g), ~round(.x, digits = 3)))

# show the table
kable(comp_formatted, caption = "Comparing numbers of sessions and rounds between the two groups",
      booktabs = T, align = "c") %>%
  scroll_box(width = "100%", height = "100%")
```

### Bet sizes

In each round, players can choose from 10 different betting sizes (from 0.25 euro till 20.00 euro). For each player, we count how often they selected each bet size, and calculated the relative proportion of each bet size being chosen. Figure \@ref(fig:bets-pattern) shows the betting patterns of both groups as heatmaps. Players were more likely to choose a small bet size rather than a large bet size, and this tendency is more pronounced in the low-risk group than in the high-risk group.

(ref:bets-pattern-cap) Heatmaps of betting patterns of all players.

```{r bets-pattern, fig.pos="center", fig.cap='(ref:bets-pattern-cap)', fig.width=5, fig.height=8}
# combine the data frame of bet counts and the median bet size
# and order the players from small median bet size to large median bet size
# for the purpose of plotting
bets <- demo %>%
  select(id_player, bet_median)

bet_sizes <- unique(bets_count$stake)
bet_sizes <- sort(bet_sizes, decreasing = FALSE)

bets_count <- bets_count %>%
  full_join(bets, by = "id_player") %>%
  arrange(group, bet_median) %>%
  mutate(
    id = c(rep(1:n_HR, each = 10), rep(1:n_LR, each = 10)),
    stake = factor(stake, levels = bet_sizes)
  )

# make a heatmap for the betting patterns of all players
bets_heatmap <- ggplot(bets_count, aes(stake, id, fill = prop)) +
  geom_tile() +
  scale_fill_gradient(low = "white", high = "black") +
  facet_wrap(~group, scales = "free_y") +
  labs(fill = "Proportion", x = "Bet (euro)") +
  theme(
    # remove player ids on the y axis
    axis.title.y=element_blank(), axis.text.y=element_blank(), axis.ticks.y=element_blank(), 
    legend.position = "top", legend.direction = "horizontal"
  )

bets_heatmap

ggsave("figure/bets_heatmap.pdf", bets_heatmap, width = 5, height = 8)
```

For each player, we calculate their (1) maximum bet size, (2) minimum bet size, (3) mean bet size, and (4) median bet size across all rounds. The high-risk group is then compared against the low-risk group on all four parameters. For all four parameters, the high-risk group bet more than the low-risk group (Table \@ref(tab:bet-compare)).

```{r bet-compare, results = "asis"}
# compare the high-risk and the low-risk group
bets_HR <- filter(demo, group == "High-Risk")
bets_LR <- filter(demo, group == "Low-Risk")

bet_max_comp    <- TES(bets_HR$bet_max, bets_LR$bet_max, paired = F, conf = 0.95)
bet_min_comp    <- TES(bets_HR$bet_min, bets_LR$bet_min, paired = F, conf = 0.95)
bet_mean_comp   <- TES(bets_HR$bet_mean, bets_LR$bet_mean, paired = F, conf = 0.95)
bet_median_comp <- TES(bets_HR$bet_median, bets_LR$bet_median, paired = F, conf = 0.95)

bet_max_comp <- bet_max_comp %>% 
  mutate(comp = "Maximum stake") %>%
  select(comp, everything())

bet_min_comp <- bet_min_comp %>% 
  mutate(comp = "Minimum stake") %>%
  select(comp, everything())

bet_mean_comp <- bet_mean_comp %>% 
  mutate(comp = "Mean stake") %>%
  select(comp, everything())

bet_median_comp <- bet_median_comp %>% 
  mutate(comp = "Median stake") %>%
  select(comp, everything())

bet_comp <- rbind(bet_max_comp, bet_min_comp, bet_mean_comp, bet_median_comp)

comp_formatted <- bet_comp %>%
  mutate(
    across(c(mean_x, sd_x, mean_y, sd_y, diff, lowerCI, upperCI), ~round(.x, digits = 2)),
    across(c(df, t, logBF), ~round(.x, digits = 1)),
    across(c(d, g), ~round(.x, digits = 3)))

# make a table
kable(comp_formatted, caption = "Comparing the maximum, minimum, mean and median bet size between two groups",
      booktabs = T, align = "c") %>%
  scroll_box(width = "100%", height = "100%")
```


### Wins

For each player, we calculate the probability of winning, and the maximum, minimum, mean and median amounts of wins on rounds where players won (the win amount is what the players received minus the stake they wagered on each round). These parameters are then compared between the two groups. Table \@ref(tab:wins-compare) shows that players in the high-risk group tended to have a **higher** probability of winning, and their winning amounts tended to be larger than those for the low-risk group.

```{r wins-compare, results = "asis"}
# compare the high-risk group and the low-risk group
wins_HR <- filter(demo, group == "High-Risk")
wins_LR <- filter(demo, group == "Low-Risk")

win_prob_comp   <- TES(wins_HR$win_prob, wins_LR$win_prob, paired = F, conf = .95)

# remove NAs as some players might not have wins in the data
wins_HR <- filter(wins_HR, !is.na(win_max))
wins_LR <- filter(wins_LR, !is.na(win_max))

win_max_comp    <- TES(wins_HR$win_max, wins_LR$win_max, paired = F, conf = .95)
win_min_comp    <- TES(wins_HR$win_min, wins_LR$win_min, paired = F, conf = .95)
win_mean_comp   <- TES(wins_HR$win_mean, wins_LR$win_mean, paired = F, conf = .95)
win_median_comp <- TES(wins_HR$win_median, wins_LR$win_median, paired = F, conf = .95)

win_prob_comp <- win_prob_comp %>%
  mutate(comp = "Win probability") %>%
  select(comp, everything())

win_max_comp <- win_max_comp %>%
  mutate(comp = "Maximum win") %>%
  select(comp, everything())

win_min_comp <- win_min_comp %>%
  mutate(comp = "Minimum win") %>%
  select(comp, everything())

win_mean_comp <- win_mean_comp %>%
  mutate(comp = "Mean win") %>%
  select(comp, everything())

win_median_comp <- win_median_comp %>%
  mutate(comp = "Median win") %>%
  select(comp, everything())

win_comp <- rbind(win_prob_comp, win_max_comp, win_min_comp, win_mean_comp, win_median_comp)

comp_formatted <- win_comp %>%
  mutate(
    across(c(mean_x, sd_x, mean_y, sd_y, diff, lowerCI, upperCI), ~round(.x, digits = 2)),
    across(c(df, t, logBF), ~round(.x, digits = 1)),
    across(c(d, g), ~round(.x, digits = 3)))

# make a table
kable(comp_formatted, caption = "Comparing the probability and amount of wins between two groups.",
      booktabs = T, align = "c") %>%
  scroll_box(width = "100%", height = "100%")
```

### Losses

On rounds where players lost, we calculate the loss (equivalent to the stake). For each player, we calculate (1) the maximum loss, (2) the minimum loss, (3) the mean loss, and (4) the median loss.

```{r losses-compare, results = "asis"}
# compare the high-risk group and the low-risk group
losses_HR <- filter(demo, group == "High-Risk")
losses_LR <- filter(demo, group == "Low-Risk")

# remove NAs as some players might not have losses in the data
losses_HR <- filter(losses_HR, !is.na(loss_max))
losses_LR <- filter(losses_LR, !is.na(loss_max))

loss_max_comp    <- TES(losses_HR$loss_max, losses_LR$loss_max, paired = F, conf = .95)
loss_min_comp    <- TES(losses_HR$loss_min, losses_LR$loss_min, paired = F, conf = .95)
loss_mean_comp    <- TES(losses_HR$loss_mean, losses_LR$loss_mean, paired = F, conf = .95)
loss_median_comp    <- TES(losses_HR$loss_median, losses_LR$loss_median, paired = F, conf = .95)

loss_max_comp <- loss_max_comp %>%
  mutate(comp = "Maximum loss") %>%
  select(comp, everything())

loss_min_comp <- loss_min_comp %>%
  mutate(comp = "Minimum loss") %>%
  select(comp, everything())

loss_mean_comp <- loss_mean_comp %>%
  mutate(comp = "Mean loss") %>%
  select(comp, everything())

loss_median_comp <- loss_median_comp %>%
  mutate(comp = "Median loss") %>%
  select(comp, everything())

loss_comp <- rbind(loss_max_comp, loss_min_comp, loss_mean_comp, loss_median_comp)

comp_formatted <- loss_comp %>%
  mutate(
    across(c(mean_x, sd_x, mean_y, sd_y, diff, lowerCI, upperCI), ~round(.x, digits = 2)),
    across(c(df, t, logBF), ~round(.x, digits = 1)),
    across(c(d, g), ~round(.x, digits = 3)))

# make a table
kable(comp_formatted, caption = "Comparing the amount of losses between two groups.",
      booktabs = T, align = "c") %>%
  scroll_box(width = "100%", height = "100%")
```

### Total spent

```{r total spent - compare, results = "asis"}
# compare the high-risk group and the low-risk group
total_spent_HR <- filter(demo, group == "High-Risk")
total_spent_LR <- filter(demo, group == "Low-Risk")

total_spent_comp    <- TES(total_spent_HR$total_spent, 
                           total_spent_LR$total_spent, 
                           paired = F, conf = .95)

total_spent_comp <- total_spent_comp %>%
  mutate(comp = "Total spent") %>%
  select(comp, everything())

comp_formatted <- total_spent_comp %>%
  mutate(
    across(c(mean_x, sd_x, mean_y, sd_y, diff, lowerCI, upperCI), ~round(.x, digits = 2)),
    across(c(df, t, logBF), ~round(.x, digits = 1)),
    across(c(d, g), ~round(.x, digits = 3)))

# make a table
kable(comp_formatted, caption = "Comparing the total amount of money spent between two groups.",
      booktabs = T, align = "c") %>%
  scroll_box(width = "100%", height = "100%")
```

## All in one table

Present all information in one table.

```{r combine-all-tables, results = "asis"}
# in the manuscript, we present all information in one table
# here we combine the separate tables into one
one_table <- rbind(count_comp, bet_comp, win_comp, loss_comp, total_spent_comp)

# for brevity, we do not report results on maximum and minimum
one_table <- one_table %>%
  filter(!comp %in% c("Maximum stake", "Minimum stake", "Maximum win",
                      "Minimum win", "Maximum loss", "Minimum loss"))

# correct the p values for multiple comparisons
one_table$p <- p.adjust(one_table$pt, method = "holm")

# do some formatting
one_table <- one_table %>%
  mutate(
    # rename comp into Parameter
    Parameter = comp,
    # for these variables, keep 1 digit after the decimal point
    across(c(mean_x:t, logBF), ~round(.x, digits = 1)),
    # for effect size g, keep 3 digits after the decimal point
    g = round(g, digits = 3),
    # report p values below .001 as < .001
    p = ifelse(p < .001, "<.001", sprintf("%.3f", round(p, digits = 3))),
    # combine mean and standard deviation for each group
    `High-Risk` = paste0(mean_x, " (", sd_x, ")"),
    `Low-Risk` = paste0(mean_y, " (", sd_y, ")")
    ) %>%
  # delete some redundant variables
  select(-c(comp:sd_y), -pt, -pw, -d, -BF) %>%
  select(Parameter, `High-Risk`, `Low-Risk`, diff, lowerCI, upperCI, df, t, p, logBF, g)

# show the table
kable(one_table, caption = "Comparing characteristics of play behavior between the two groups.",
      booktabs = T, align = "c") %>%
  scroll_box(width = "100%", height = "100%")

# show the table in latex code
# kable(one_table, "latex",
#       caption = "Comparing characteristics of play behavior between the two groups.",
#       booktabs = T, align = "l") %>%
#   kable_styling(latex_options = "scale_down")
```
