==> Complete/code/analysis/05-change_in_stake.Rmd <==
# Change in stake

```{r stake-change-load-data}
stake_change <- read_csv("../../data/processed/stake_change.csv")
```

In this section, we analyze change in stake after a win and a loss. For this analysis, players need to have at least 5 rounds following a win and 5 rounds following a loss to be included. The table below shows the number of players and rounds left in this analysis.

```{r stake change - data preparation, results='asis'}
# only participants with at least 5 rounds following wins and losses
# are included in the plot and following analyses
players_incldue <- stake_change %>%
  filter(n_round >= 5) %>%
  pivot_wider(
    id_cols = c(group, id_player), names_from = prev_outcome,
    values_from = stake_change_prop
  ) %>%
  drop_na() %>%
  .$id_player

stake_change <- filter(stake_change, id_player %in% players_incldue)

# calculate the remaining number of players and rounds
count <- stake_change %>%
  group_by(group, id_player) %>%
  summarize(total_count = sum(n_round)) %>%
  group_by(group) %>%
  summarize(
    player = n(),
    round = mean(total_count),
    sd = sd(total_count),
    min = min(total_count),
    max = max(total_count)
  )

kable(count, caption = "Number of players and rounds left")

```


## Plot

For each player, we calculate (1) the probability (%) of changing stake after a win and a loss; (2) the average change in stake amount (euro) after a win and a loss. Since players can only choose from 10 amount levels, as a second metric of stake change, we convert stake amount into stake level (with 1 being the lowest stake level, i.e., 0.25 euro, and 10 being the highest stake level, i.e., 20.00 euro) and conduct the same analysis on the change in stake level.

```{r stake-change-prob-plot}
# plot the probability of changing stake after a win and a loss
stake_change_prob_summary <- stake_change %>%
  group_by(group) %>%
  group_modify(~Rmisc::summarySEwithin(.x, measurevar = "stake_change_prop", idvar = "id_player",
                        withinvars = "prev_outcome"))

stake_change_prob_plot <-
  ggplot(stake_change_prob_summary, aes(prev_outcome, stake_change_prop, color = group)) +
  geom_point(position = position_dodge(width = 0.5)) +
  geom_line(aes(group = group), position = position_dodge(width = 0.5), linetype = "dashed") +
  geom_errorbar(aes(ymin = stake_change_prop-ci, ymax = stake_change_prop+ci), 
                width =0.2, position = position_dodge(width = 0.5)) +
  # use colorblind-friendly color
  scale_color_manual(values = color_values) + 
  scale_x_discrete(labels=c("loss" = "Loss", "win" = "Win")) +
  labs(x = "Previous outcome", 
       y = "Probability of changing stake (%)", color = "Group") +
  theme(legend.position = "top", legend.direction = "horizontal")
```


```{r stake-change-size-plot}
# the mean change in stake size (euro) after a win and a loss
stake_change_size_summary <- stake_change %>%
  group_by(group) %>%
  group_modify(~Rmisc::summarySEwithin(.x, 
                                       measurevar = "stake_change_size_overall", 
                                       idvar = "id_player",
                                       withinvars = "prev_outcome"))

stake_change_size_plot <- stake_change_size_summary %>%
  ggplot(aes(prev_outcome, stake_change_size_overall, color = group)) +
  geom_point(position = position_dodge(width = 0.5)) +
  geom_line(aes(group = group), position = position_dodge(width = 0.5), linetype = "dashed") +
  geom_errorbar(aes(ymin = stake_change_size_overall-ci, ymax = stake_change_size_overall+ci), 
                width =0.2, position = position_dodge(width = 0.5)) +
  # use colorblind-friendly color
  scale_color_manual(values = color_values) + 
  scale_x_discrete(labels=c("loss" = "Loss", "win" = "Win")) +
  labs(x = "Previous outcome", 
       y = "Change in stake size (euro)", color = "Group") +
  theme(legend.position = "top", legend.direction = "horizontal")
```


```{r stake-change-level-plot}
# the mean change in stake level after a win and a loss
stake_change_level_summary <- stake_change %>%
  group_by(group) %>%
  group_modify(~Rmisc::summarySEwithin(.x, measurevar = "stake_change_level_overall", 
                                       idvar = "id_player",
                                       withinvars = "prev_outcome"))

stake_change_level_plot <- stake_change_level_summary %>%
  ggplot(aes(prev_outcome, stake_change_level_overall, color = group)) +
  geom_point(position = position_dodge(width = 0.5)) +
  geom_line(aes(group = group), position = position_dodge(width = 0.5), linetype = "dashed") +
  geom_errorbar(aes(ymin = stake_change_level_overall-ci, ymax = stake_change_level_overall+ci), 
                width =0.2, position = position_dodge(width = 0.5)) +
  # use colorblind-friendly color
  scale_color_manual(values = color_values) + 
  scale_x_discrete(labels=c("loss" = "Loss", "win" = "Win")) +
  labs(x = "Previous outcome", 
       y = "Change in stake level", color = "Group") +
  theme(legend.position = "top", legend.direction = "horizontal")
```

(ref:stake-change-plot-cap) The probability of changing stake (left),  the average change in stake amount (euro; middle) and the average change in stake level (right) after a win and a loss for both groups. Error bars stand for 95\% within-subject confidence intervals.

```{r stake-change-plot, fig.cap='(ref:stake-change-plot-cap)', fig.width=9, fig.height=3.5}
# combine three plots
stake_change_plot <- ggarrange(stake_change_prob_plot,
                               stake_change_size_plot,
                               stake_change_level_plot,  
                               nrow = 1, common.legend = TRUE, legend="top")

stake_change_plot

ggsave("figure/stake_change.pdf", stake_change_plot, width = 9, height = 3.5)
```

## ANOVAs

We conduct three ANOVAs for the three dependent variables, using prior game outcome (loss vs. win, within subjects) and risk level (high vs. low, between subjects) as independent variables.

```{r stake-change-anova, results = "asis", cache=TRUE}
# run ANOVAs
stake_change_prob_anova <-
  aov_ez(id = "id_player"
         , dv = "stake_change_prop"
         , data = stake_change
         , between = "group"
         , within = "prev_outcome")

stake_change_prob_anova <- nice(stake_change_prob_anova)

stake_change_size_anova <-
  aov_ez(id = "id_player"
         , dv = "stake_change_size_overall"
         , data = stake_change
         , between = "group"
         , within = "prev_outcome")

stake_change_size_anova <- nice(stake_change_size_anova)

stake_change_level_anova <-
  aov_ez(id = "id_player"
         , dv = "stake_change_level_overall"
         , data = stake_change
         , between = "group"
         , within = "prev_outcome")

stake_change_level_anova <- nice(stake_change_level_anova)

# combine all anova tables and do some formatting
change_anova_table <- 
  bind_rows(
    stake_change_prob_anova,
    stake_change_size_anova,
    stake_change_level_anova
  ) %>%
  mutate(
    Parameter = rep(c("Prob of change (%)", "Change in amount (euro)", "Change in level"),
                    each = 3),
    Effect = rep(c("Risk Level", "Prior Outcome", "Interaction"), 3),
    # remove asterisks from the F values
    `F` = str_remove_all(`F`, "[*]"),
    # remove whitespace
    `F` = str_trim(`F`),
    p = `p.value`
  ) %>%
  select(Parameter, Effect, df, MSE, `F`, ges, p)

# show the table
kable(change_anova_table, 
      caption = "Effects of group and prior outcome on change in stake",
      booktabs = T, align = "l") %>%
  collapse_rows(columns = 1) %>%
  kable_styling(latex_options = "scale_down")

# # show the table in latex code
# kable(change_anova_table, "latex",
#       caption = "Effects of group and prior outcome on change in stake",
#       booktabs = T, align = "l") %>%
#   collapse_rows(columns = 1) %>%
#   kable_styling(latex_options = "scale_down")
```

## Pairwise comparisons

Next we conduct a series of pairwise comparisons for each of the three dependent variables. We first compare the effect of prior outcome on the probability of changing stake within each group (two comparisons). We then compare the probability of changing stake after a loss between two groups, and the probability of changing stake after a win between two groups. We further calculate a difference score between loss and win for each player, and compare the difference scores between groups as well. The same comparisons are conducted for the change in stake amount and the change in stake level. In total we conduct 15 comparisons. P values are corrected for multiple comparisons using the Holm-Bonferroni method.

```{r change in stake pairwise comparisons, results='asis'}
#### Probability of changing stake ####

# descriptive statistics
stake_change_per_outcome <- stake_change %>%
  group_by(prev_outcome) %>%
  summarize(
    mean = mean(stake_change_prop),
    sd = sd(stake_change_prop)
    )
  
stake_change_per_cell <- stake_change %>%
  group_by(group, prev_outcome) %>%
  summarize(
    mean = mean(stake_change_prop),
    sd = sd(stake_change_prop)
    )
  
# within group comparisons on prob of change
stake_change_prob <- stake_change %>%
  pivot_wider(id_cols = c(group, id_player),
              names_from = prev_outcome,
              values_from = stake_change_prop) %>%
  mutate(diff = loss - win)

change_prob_within <- stake_change_prob %>%
  group_by(group) %>%
  group_modify(~TES(.x$loss, .x$win, paired = TRUE)) %>%
  mutate(d = dz, g = gav) %>%
  ungroup() %>% select(-group, -dz, -gav, -dav)

# between group comparisons on prob of change
change_prob_between1 <- TES(filter(stake_change_prob, group == "High-Risk")$loss,
                            filter(stake_change_prob, group == "Low-Risk")$loss, paired = FALSE)

change_prob_between2 <- TES(filter(stake_change_prob, group == "High-Risk")$win,
                            filter(stake_change_prob, group == "Low-Risk")$win, paired = FALSE)

change_prob_between3 <- TES(filter(stake_change_prob, group == "High-Risk")$diff,
                            filter(stake_change_prob, group == "Low-Risk")$diff, paired = FALSE)

change_prob_between <- bind_rows(change_prob_between1, change_prob_between2, change_prob_between3)

# combine
change_prob_comp <- bind_rows(change_prob_within, change_prob_between)

#### Change in stake size (euro) ####

# descriptive statistics
stake_size_per_outcome <- stake_change %>%
  group_by(prev_outcome) %>%
  summarize(
    mean = mean(stake_change_size_overall) * 100, # turn into cents
    sd = sd(stake_change_size_overall) * 100
    )
  
stake_size_per_cell <- stake_change %>%
  group_by(group, prev_outcome) %>%
  summarize(
    mean = mean(stake_change_size_overall) * 100,
    sd = sd(stake_change_size_overall) * 100
    )
  
# within group comparisons on change in stake size (euro)
stake_change_size <- stake_change %>%
  pivot_wider(id_cols = c(group, id_player),
              names_from = prev_outcome,
              values_from = stake_change_size_overall) %>%
  mutate(diff = loss - win)

change_size_within <- stake_change_size %>%
  group_by(group) %>%
  group_modify(~TES(.x$loss, .x$win, paired = TRUE)) %>%
  mutate(d = dz, g = gav) %>%
  ungroup() %>% select(-group, -dz, -gav, -dav)

# between group comparisons on change in stake size (euro)
stake_change_size <- stake_change_size %>%
  mutate(diff = loss - win)

change_size_between1 <- TES(filter(stake_change_size, group == "High-Risk")$loss,
                            filter(stake_change_size, group == "Low-Risk")$loss, paired = FALSE)

change_size_between2 <- TES(filter(stake_change_size, group == "High-Risk")$win,
                            filter(stake_change_size, group == "Low-Risk")$win, paired = FALSE)

change_size_between3 <- TES(filter(stake_change_size, group == "High-Risk")$diff,
                            filter(stake_change_size, group == "Low-Risk")$diff, paired = FALSE)

change_size_between <- bind_rows(change_size_between1, change_size_between2, change_size_between3)

# combine
change_size_comp <- bind_rows(change_size_within, change_size_between)


#### Change in stake level ####

# descriptive statistics
stake_level_per_outcome <- stake_change %>%
  group_by(prev_outcome) %>%
  summarize(
    mean = mean(stake_change_level_overall), 
    sd = sd(stake_change_level_overall)
    )
  
stake_level_per_cell <- stake_change %>%
  group_by(group, prev_outcome) %>%
  summarize(
    mean = mean(stake_change_level_overall), 
    sd = sd(stake_change_level_overall)
    )

# within group comparisons on change in stake level
stake_change_level <- stake_change %>%
  pivot_wider(id_cols = c(group, id_player),
              names_from = prev_outcome,
              values_from = stake_change_level_overall) %>%
  mutate(diff = loss - win)

change_level_within <- stake_change_level %>%
  group_by(group) %>%
  group_modify(~TES(.x$loss, .x$win, paired = TRUE)) %>%
  mutate(d = dz, g = gav) %>%
  ungroup() %>% select(-group, -dz, -gav, -dav)

# between group comparisons on change in stake level

change_level_between1 <- TES(filter(stake_change_level, group == "High-Risk")$loss,
                            filter(stake_change_level, group == "Low-Risk")$loss, paired = FALSE)

change_level_between2 <- TES(filter(stake_change_level, group == "High-Risk")$win,
                            filter(stake_change_level, group == "Low-Risk")$win, paired = FALSE)

change_level_between3 <- TES(filter(stake_change_level, group == "High-Risk")$diff,
                            filter(stake_change_level, group == "Low-Risk")$diff, paired = FALSE)

change_level_between <- bind_rows(change_level_between1, 
                                  change_level_between2, 
                                  change_level_between3)

# combine
change_level_comp <- bind_rows(change_level_within, change_level_between)


#### Combine all comparisons ####

# combine all results and correct p values for multiple comparisons
comp <- bind_rows(change_prob_comp, change_size_comp, change_level_comp) %>%
  mutate(
    Parameter = rep(c("Prob of change (%)", "Change in amount (euro)", "Change in level"),
                    each = 5),
    Comparison = rep(c("Loss vs. Win (High-Risk)", "Loss vs. Win (Low-Risk)", 
                       "High- vs. Low-Risk (Loss)", "High- vs. Low-Risk (Win)",
                       "High- vs. Low-Risk (Difference)"),
                     3),
    pt = p.adjust(pt, method = "holm")
  ) 

# do some formatting
comp <- comp %>%
  mutate(
    diff = ifelse(Parameter == "Prob of change (%)", 
                  sprintf("%.2f", round(diff, digits = 2)), 
                  sprintf("%.4f", round(diff, digits = 4))),
    lowerCI = ifelse(Parameter == "Prob of change (%)", 
                  sprintf("%.2f", round(lowerCI, digits = 2)), 
                  sprintf("%.4f", round(lowerCI, digits = 4))),
    upperCI = ifelse(Parameter == "Prob of change (%)", 
                  sprintf("%.2f", round(upperCI, digits = 2)), 
                  sprintf("%.4f", round(upperCI, digits = 4))),
    across(c(d, g), ~round(.x, digits = 3)),
    t = round(t, digits = 1),
    df = round(df, digits = 1),
    p = ifelse(pt < .001, "<.001", sprintf("%.3f", round(pt, digits = 3))),
    logBF = round(logBF, digits = 2)
  ) %>%
  select(Parameter, Comparison, diff:t, p, logBF, g) 

# show the table
kable(comp, 
      caption = "Pairwise comparisons for effects on change in stake",
      booktabs = T, align = "l") %>%
  collapse_rows(columns = 1) %>%
  kable_styling(latex_options = "scale_down") %>%
  scroll_box(width = "100%", height = "100%")

# # show the table in latex code
# kable(comp, "latex",
#       caption = "Pairwise comparisons for effects on change in stake",
#       booktabs = T, align = "l") %>%
#   collapse_rows(columns = 1) %>%
#   kable_styling(latex_options = "scale_down")
```
