==> Complete/code/analysis/06-speed_of_play.Rmd <==
# Speed of play


```{r speed of play load data}
rt <- read_csv("../../data/processed/rt.csv")
```

In this section, we analyze the speed of play after a win or a loss (i.e., from starting the current game till putting in the first column). We do not use the RT of starting a game, because in this data set, the 'startRT' is from putting in the last column in the previous game till starting the current game. The 'startRT' therefore includes the visual and auditory feedback players receive (which differs between winning and losing), as well as the time potentially spent on adjusting the stake for the current game. The RT of putting in the first column, on the other hand, is not influenced by these factors. 

Players need to have at least 5 rounds following a win and 5 rounds following a loss to be included. The table below shows the number of players and rounds left in this analysis.

```{r speed of play - data preparation, results='asis'}
# only participants with at least 5 rounds following wins and losses
# are included in the plot and following analyses
players_incldue <- rt %>%
  filter(n_round >= 5) %>%
  pivot_wider(
    id_cols = c(group, id_player), names_from = prev_outcome,
    values_from = rt_mean
  ) %>%
  drop_na() %>%
  .$id_player

rt <- filter(rt, id_player %in% players_incldue)

# calculate the remaining number of players and rounds
count <- rt %>%
  group_by(group, id_player) %>%
  summarize(total_count = sum(n_round)) %>%
  group_by(group) %>%
  summarize(
    player = n(),
    round = mean(total_count),
    sd = sd(total_count),
    min = min(total_count),
    max = max(total_count)
  )

kable(count, caption = "Number of players and rounds left")

```


## Plot

(ref:rt-plot-cap) Response time (left: raw RT; right: RT z scores) after a win and after a loss. 

```{r rt-overall-plot, fig.cap="(ref:rt-plot-cap)", fig.width=5, fig.height=3}
# for each group, calculate the mean RT after winning and losing
rt_summary <- rt %>%
  group_by(group) %>%
  group_modify(~Rmisc::summarySEwithin(.x, measurevar = "rt_mean", idvar = "id_player",
                        withinvars = "prev_outcome"))

# plot the results
rt_plot <- rt_summary %>%
  ggplot(aes(prev_outcome, rt_mean, color = group)) +
  geom_point(position = position_dodge(width = 0.5)) +
  geom_line(aes(group = group), position = position_dodge(width = 0.5), linetype = "dashed") +
  geom_errorbar(aes(ymin = rt_mean-ci, ymax = rt_mean+ci), width = 0.2, 
                position = position_dodge(width = 0.5)) +
  # use colorblind-friendly color
  scale_color_manual(values = color_values) + 
  scale_x_discrete(labels=c("loss" = "Loss", "win" = "Win")) +
  labs(x = "Previous outcome", y = "Response time (milliseconds)", color = "Group") 


# for each group, calculate the mean RT z score after winning and losing
rt_z_summary <- rt %>%
  group_by(group) %>%
  group_modify(~Rmisc::summarySEwithin(.x, measurevar = "rt_z_mean", idvar = "id_player",
                        withinvars = "prev_outcome"))

rt_z_plot <- rt_z_summary %>%
  ggplot(aes(prev_outcome, rt_z_mean, color = group)) +
  geom_point(position = position_dodge(width = 0.5)) +
  geom_line(aes(group = group), position = position_dodge(width = 0.5), linetype = "dashed") +
  geom_errorbar(aes(ymin = rt_z_mean-ci, ymax = rt_z_mean+ci), width =0.2, 
                position = position_dodge(width = 0.5)) +
  # use colorblind-friendly color
  scale_color_manual(values = color_values) + 
  scale_x_discrete(labels=c("loss" = "Loss", "win" = "Win")) +
  labs(x = "Previous outcome", y = "Response time (z score)", color = "Group") 

# combine the two plots
rt_combined_plot <- ggarrange(rt_plot, rt_z_plot,
                               nrow = 1, common.legend = TRUE, legend="top")

rt_combined_plot

ggsave("figure/rt.pdf", rt_combined_plot, width = 5, height = 3)


```

## ANOVAs

For both the raw RTs and RT z scores, we conduct a mixed ANOVA, with prior game outcome (loss vs. win, within subjects) and player risk level (high vs. low, between subjects) as independent variables.

```{r rt-analysis, cache=TRUE, results="asis"}
# run ANOVAs
rt_anova <-
  afex::aov_ez(id = "id_player"
               , dv = "rt_mean"
               , data = rt
               , between = "group"
               , within = "prev_outcome")

rt_anova <- nice(rt_anova)

rt_z_anova <-
  afex::aov_ez(id = "id_player"
               , dv = "rt_z_mean"
               , data = rt
               , between = "group"
               , within = "prev_outcome")

rt_z_anova <- nice(rt_z_anova)

# combine anova tables and do some formatting
rt_anova_table <- 
  bind_rows(
    rt_anova,
    rt_z_anova,
  ) %>%
  mutate(
    Parameter = rep(c("RT (milliseconds)", "RT (z scores)"),
                    each = 3),
    Effect = rep(c("Risk Level", "Prior Outcome", "Interaction"), 2),
    # remove asterisks from the F values
    `F` = str_remove_all(`F`, "[*]"),
    # remove whitespace
    `F` = str_trim(`F`),
    p = `p.value`
  ) %>%
  select(Parameter, Effect, df, MSE, `F`, ges, p)
  
# show the table
kable(rt_anova_table, 
      caption = "Effects of group and prior outcome on speed of play",
      booktabs = T, align = "l") %>%
  collapse_rows(columns = 1) %>%
  kable_styling(latex_options = "scale_down")

# # show the table in latex code
# kable(rt_anova_table, "latex",
#       caption = "Effects of group and prior outcome on speed of play",
#       booktabs = T, align = "l") %>%
#   collapse_rows(columns = 1) %>%
#   kable_styling(latex_options = "scale_down")
```

## Pairwise comparisons

We similarly conduct a series of within- and between-group pairwise comparisons. More specifically, we compare the effect of a loss vs. a win on the speed of play within each group. We then compare the speed of play after a loss, after a loss and the difference in speed of play between loss and win, between the two groups. In total we have 10 comparisons - p values are again corrected for multiple comparisons using the Holm-Bonferroni method.

```{r rt pairwise comparisons, results='asis'}
#### Raw RTs ####

# descriptive statistics
rt_group <- rt %>%
  group_by(group) %>%
  summarize(
    mean = mean(rt_mean),
    sd = sd(rt_mean)
  )

rt_outcome <- rt %>%
  group_by(prev_outcome) %>%
  summarize(
    mean = mean(rt_mean),
    sd = sd(rt_mean)
  )

rt_all_cells <- rt %>%
  group_by(group, prev_outcome) %>%
  summarize(
    mean = mean(rt_mean),
    sd = sd(rt_mean)
  )

# pairwise comparisons
rt_wide <- rt %>%
  pivot_wider(id_cols = c(group, id_player), 
              names_from = prev_outcome, values_from = rt_mean) %>%
  mutate(diff = loss - win)

# first, compare the effect of loss vs. win within each group
rt_comp_within <-  rt_wide %>%
  group_by(group) %>%
  group_modify(~TES(.x$loss, .x$win, paired = TRUE)) %>%
  mutate(d = dz, g = gav) %>%
  ungroup() %>% select(-group, -dz, -gav, -dav)

# next, compare the two groups
comp_between1 <- TES(filter(rt_wide, group == "High-Risk")$loss,
                     filter(rt_wide, group == "Low-Risk")$loss, paired = FALSE)

comp_between2 <- TES(filter(rt_wide, group == "High-Risk")$win,
                     filter(rt_wide, group == "Low-Risk")$win, paired = FALSE)

comp_between3 <- TES(filter(rt_wide, group == "High-Risk")$diff,
                     filter(rt_wide, group == "Low-Risk")$diff, paired = FALSE)

rt_comp_between <- bind_rows(comp_between1, comp_between2, comp_between3)

#### RT z scores ####

# descriptive statistics
rt_z_group <- rt %>%
  group_by(group) %>%
  summarize(
    mean = mean(rt_z_mean),
    sd = sd(rt_z_mean)
  )

rt_z_outcome <- rt %>%
  group_by(prev_outcome) %>%
  summarize(
    mean = mean(rt_z_mean),
    sd = sd(rt_z_mean)
  )

rt_z_all_cells <- rt %>%
  group_by(group, prev_outcome) %>%
  summarize(
    mean = mean(rt_z_mean),
    sd = sd(rt_z_mean)
  )

# pairwise comparisons
rt_z_wide <- rt %>%
  pivot_wider(id_cols = c(group, id_player), 
              names_from = prev_outcome, values_from = rt_z_mean) %>%
  mutate(diff = loss - win)

# first, compare the effect of loss vs. win within each group
rt_z_comp_within <-  rt_z_wide %>%
  group_by(group) %>%
  group_modify(~TES(.x$loss, .x$win, paired = TRUE)) %>%
  mutate(d = dz, g = gav) %>%
  ungroup() %>% select(-group, -dz, -gav, -dav)

# next, compare the two groups
comp_between1 <- TES(filter(rt_z_wide, group == "High-Risk")$loss,
                     filter(rt_z_wide, group == "Low-Risk")$loss, paired = FALSE)

comp_between2 <- TES(filter(rt_z_wide, group == "High-Risk")$win,
                     filter(rt_z_wide, group == "Low-Risk")$win, paired = FALSE)

comp_between3 <- TES(filter(rt_z_wide, group == "High-Risk")$diff,
                     filter(rt_z_wide, group == "Low-Risk")$diff, paired = FALSE)

rt_z_comp_between <- bind_rows(comp_between1, comp_between2, comp_between3)

#### Combine all comparisons ####

# combine all results and correct p values for multiple comparisons
comp <- 
  bind_rows(
    rt_comp_within, 
    rt_comp_between, 
    rt_z_comp_within,
    rt_z_comp_between
  ) %>%
  mutate(
    Parameter = rep(c("RT (milliseconds)", "RT (z scores)"),
                    each = 5),
    Comparison = rep(c("Loss vs. Win (High-Risk)", "Loss vs. Win (Low-Risk)", 
                       "High- vs. Low-Risk (Loss)", "High- vs. Low-Risk (Win)",
                       "High- vs. Low-Risk (Difference)"),
                     2),
    pt = p.adjust(pt, method = "holm")
  ) 

# do some formatting
comp <- comp %>%
  mutate(
    diff = ifelse(Parameter == "RT (milliseconds)", 
                  sprintf("%.1f", round(diff, digits = 1)), 
                  sprintf("%.3f", round(diff, digits = 3))),
    lowerCI = ifelse(Parameter == "RT (milliseconds)", 
                  sprintf("%.1f", round(lowerCI, digits = 1)), 
                  sprintf("%.3f", round(lowerCI, digits = 3))),
    upperCI = ifelse(Parameter == "RT (milliseconds)", 
                  sprintf("%.1f", round(upperCI, digits = 1)), 
                  sprintf("%.3f", round(upperCI, digits = 3))),
    across(c(d, g), ~round(.x, digits = 3)),
    t = round(t, digits = 1),
    df = round(df, digits = 1),
    p = ifelse(pt < .001, "<.001", sprintf("%.3f", round(pt, digits = 3))),
    logBF = round(logBF, digits = 2)
  ) %>%
  select(Parameter, Comparison, diff:t, p, logBF, g) 

# show the table
kable(comp, 
      caption = "Pairwise comparisons for effects on speed of play",
      booktabs = T, align = "l") %>%
  collapse_rows(columns = 1) %>%
  kable_styling(latex_options = "scale_down") %>%
  scroll_box(width = "100%", height = "100%")


# kable(comp, "latex",
#       caption = "Pairwise comparisons for effects on speed of play",
#       booktabs = T, align = "l") %>%
#   collapse_rows(columns = 1) %>%
#   kable_styling(latex_options = "scale_down")
```
